requirements.txt
Flask
pandas
pandasql
tk
pyqt5
groq
python-dotenv